Restructure modules related to training setup and monitoring
