+++ corebehrt/configs/pretrain.yaml
@@ -61,11 +61,12 @@ scheduler:
 
 metrics:
   top1:
-    _target_: corebehrt.evaluation.metrics.PrecisionAtK
+    _target_: corebehrt.modules.monitoring.metrics.PrecisionAtK
     topk: 1
   top10:
-    _target_: corebehrt.evaluation.metrics.PrecisionAtK
+    _target_: corebehrt.modules.monitoring.metrics.PrecisionAtK
+
     topk: 10
   mlm_loss:
-    _target_: corebehrt.evaluation.metrics.LossAccessor
+    _target_: corebehrt.modules.monitoring.metrics.LossAccessor
     loss_name: loss
