+++ corebehrt/main/finetune_cv.py
@@ -8,8 +8,11 @@ from corebehrt.common.config import load_config
 from corebehrt.common.initialize import ModelManager
 from corebehrt.common.loader import load_and_select_splits
 from corebehrt.common.setup import DirectoryPreparer, get_args
-from corebehrt.evaluation.utils import compute_and_save_scores_mean_std
+from corebehrt.modules.monitoring.metric_aggregation import (
+    compute_and_save_scores_mean_std,
+)
 from corebehrt.functional.load import load_pids
+
 from corebehrt.functional.split import (
     get_n_splits_cv_pids,
     split_into_test_and_train_val_pids,
