+++ corebehrt/modules/monitoring/metric_aggregation.py
@@ -1,14 +1,11 @@
 import os
+from datetime import datetime
+from os.path import join
 
 import numpy as np
+import pandas as pd
 import torch
 from sklearn.metrics import precision_recall_curve, roc_curve
-from torch.utils.data import DataLoader
-from tqdm import tqdm
-
-
-def get_tqdm(dataloader: DataLoader) -> tqdm:
-    return tqdm(dataloader, total=len(dataloader), mininterval=10)
 
 
 def compute_avg_metrics(metric_values: dict):
@@ -65,3 +62,30 @@ def save_metrics_to_csv(run_folder: str, metrics: dict, epoch: int, mode="val") 
         file.write("metric,value\n")
         for key, value in metrics.items():
             file.write(f"{key},{value}\n")
+
+
+def compute_and_save_scores_mean_std(
+    n_splits: int, finetune_folder: str, mode="val"
+) -> None:
+    """Compute mean and std of test/val scores. And save to finetune folder."""
+    scores = []
+    for fold in range(1, n_splits + 1):
+        fold_checkpoints_folder = join(finetune_folder, f"fold_{fold}", "checkpoints")
+        last_epoch = max(
+            [
+                int(f.split("_")[-2].split("epoch")[-1])
+                for f in os.listdir(fold_checkpoints_folder)
+                if f.startswith("checkpoint_epoch")
+            ]
+        )
+        table_path = join(fold_checkpoints_folder, f"{mode}_scores_{last_epoch}.csv")
+        if not os.path.exists(table_path):
+            continue
+        fold_scores = pd.read_csv(
+            join(fold_checkpoints_folder, f"{mode}_scores_{last_epoch}.csv")
+        )
+        scores.append(fold_scores)
+    scores = pd.concat(scores)
+    scores_mean_std = scores.groupby("metric")["value"].agg(["mean", "std"])
+    date = datetime.now().strftime("%Y%m%d-%H%M")
+    scores_mean_std.to_csv(join(finetune_folder, f"{mode}_scores_mean_std_{date}.csv"))
